finished

+++ database/migrations/2020_05_07_125100_create_students_table.php
@@ -16,7 +16,7 @@ class CreateStudentsTable extends Migration
         Schema::create('students', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id');
-            $table->integer('registration_number');
+            $table->bigInteger('registration_number');
             $table->string('name');
             $table->integer('gender');
             $table->string('phone_number');
